# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GeneSilico EHR API",
        "summary": "Ontada (iKnowMed) FHIR + Stedi eligibility for Texas Oncology.",
...
        "contact": {
            "name": "GeneSilico",
            "email": "developers@genesilico.ai"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8090",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/api/casebooks": {
            "get": {
                "tags": [
                    "casebooks"
                ],
                "summary": "List Casebooks",
                "operationId": "list_casebooks_api_casebooks_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasebookList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "casebooks"
                ],
                "summary": "Create Casebook",
                "operationId": "create_casebook_api_casebooks_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CasebookIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasebookOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/casebooks/{casebook_id}": {
            "get": {
                "tags": [
                    "casebooks"
                ],
                "summary": "Get Casebook",
                "operationId": "get_casebook_api_casebooks__casebook_id__get",
                "parameters": [
                    {
                        "name": "casebook_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Casebook Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasebookOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/casebooks/{casebook_id}/packages": {
            "get": {
                "tags": [
                    "casebooks"
                ],
                "summary": "Casebook Packages",
                "description": "Pre-auth packages belonging to this patient, newest first.",
                "operationId": "casebook_packages_api_casebooks__casebook_id__packages_get",
                "parameters": [
                    {
                        "name": "casebook_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Casebook Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasebookPackageList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/casebooks/{casebook_id}/discover-coverage": {
            "post": {
                "tags": [
                    "casebooks"
                ],
                "summary": "Discover Coverage",
                "description": "Find every policy this patient is covered under.\n\nAnswers a question the chart usually cannot: is the payer we are about to\nbill actually the primary, and is the patient a dependent on a plan nobody\nrecorded? Getting that order wrong is a denial.",
                "operationId": "discover_coverage_api_casebooks__casebook_id__discover_coverage_post",
                "parameters": [
                    {
                        "name": "casebook_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Casebook Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoverageDiscovery"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "428": {
                        "description": "A precondition is unmet \u2014 typically Ontada is not connected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "501": {
                        "description": "Deliberately not implemented; the reason is in the message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "An upstream (Stedi, Ontada) failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "A capability is not configured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CasebookIn": {
                "properties": {
                    "patient_name": {
                        "type": "string",
                        "minLength": 1,
                        "title": "Patient Name"
                    },
                    "mrn": {
                        "type": "string",
                        "title": "Mrn",
                        "default": ""
                    },
                    "birth_date": {
                        "type": "string",
                        "title": "Birth Date",
                        "default": ""
                    },
                    "gender": {
                        "type": "string",
                        "title": "Gender",
                        "default": ""
                    },
                    "primary_diagnosis": {
                        "type": "string",
                        "title": "Primary Diagnosis",
                        "default": ""
                    },
                    "diagnosis_code": {
                        "type": "string",
                        "title": "Diagnosis Code",
                        "default": ""
                    },
                    "stage": {
                        "type": "string",
                        "title": "Stage",
                        "default": ""
                    },
                    "cancer_center_id": {
                        "type": "string",
                        "title": "Cancer Center Id",
                        "default": ""
                    },
                    "hospital_id": {
                        "type": "string",
                        "title": "Hospital Id",
                        "default": ""
                    },
                    "oncologist": {
                        "type": "string",
                        "title": "Oncologist",
                        "default": ""
                    },
                    "ontada_fhir_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ontada Fhir Id"
                    },
                    "fhir_snapshot": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Fhir Snapshot"
                    }
                },
                "type": "object",
                "required": [
                    "patient_name"
                ],
                "title": "CasebookIn"
            },
            "CasebookList": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/CasebookOut"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "title": "CasebookList"
            },
            "CasebookOut": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "patientName": {
                        "type": "string",
                        "title": "Patientname"
                    },
                    "mrn": {
                        "type": "string",
                        "title": "Mrn"
                    },
                    "birthDate": {
                        "type": "string",
                        "title": "Birthdate"
                    },
                    "gender": {
                        "type": "string",
                        "title": "Gender"
                    },
                    "primaryDiagnosis": {
                        "type": "string",
                        "title": "Primarydiagnosis"
                    },
                    "diagnosisCode": {
                        "type": "string",
                        "title": "Diagnosiscode"
                    },
                    "stage": {
                        "type": "string",
                        "title": "Stage"
                    },
                    "cancerCenterId": {
                        "type": "string",
                        "title": "Cancercenterid"
                    },
                    "hospitalId": {
                        "type": "string",
                        "title": "Hospitalid"
                    },
                    "oncologist": {
                        "type": "string",
                        "title": "Oncologist"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "ontadaFhirId": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ontadafhirid"
                    },
                    "resourceCounts": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Resourcecounts"
                    },
                    "cancerAiConfidence": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Canceraiconfidence",
                        "description": "Always null \u2014 gSage is not wired to this module"
                    },
                    "dtEfficacyScore": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Dtefficacyscore",
                        "description": "Always null \u2014 gSage is not wired to this module"
                    },
                    "createdAt": {
                        "type": "string",
                        "title": "Createdat"
                    },
                    "updatedAt": {
                        "type": "string",
                        "title": "Updatedat"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "patientName",
                    "mrn",
                    "birthDate",
                    "gender",
                    "primaryDiagnosis",
                    "diagnosisCode",
                    "stage",
                    "cancerCenterId",
                    "hospitalId",
                    "oncologist",
                    "status",
                    "ontadaFhirId",
                    "resourceCounts",
                    "createdAt",
                    "updatedAt"
                ],
                "title": "CasebookOut"
            },
            "CasebookPackageList": {
                "properties": {
                    "count": {
                        "type": "integer",
                        "title": "Count"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/CasebookPackageRow"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "title": "CasebookPackageList"
            },
...
            "ErrorEnvelope": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorEnvelope",
                "description": "Every non-2xx response from this API has exactly this shape."
            },
...
        }
    },
...
}